minor updates to C code

+++ pyModeS/c_decoder/bds/bds05.pyx
@@ -13,14 +13,14 @@ from libc.math cimport NAN as nan
 
 
 @cython.cdivision(True)
-def airborne_position(bytes msg0 not None, bytes msg1 not None, long t0, long t1):
+def airborne_position(bytes msg0 not None, bytes msg1 not None, double t0, double t1):
     """Decode airborn position from a pair of even and odd position message
 
     Args:
         msg0 (string): even message (28 bytes hexadecimal string)
         msg1 (string): odd message (28 bytes hexadecimal string)
-        t0 (int): timestamps for the even message
-        t1 (int): timestamps for the odd message
+        t0 (double): timestamps for the even message
+        t1 (double): timestamps for the odd message
 
     Returns:
         (float, float): (latitude, longitude) of the aircraft
